+++ examples/nextjs/src/app/providers.tsx
@@ -24,7 +24,8 @@ const walletManager = new WalletManager({
     {
       id: WalletId.MAGIC,
       options: { apiKey: 'pk_live_D17FD8D89621B5F3' }
-    }
+    },
+    WalletId.MNEMONIC
   ],
   network: NetworkId.TESTNET
 })