+++ packages/use-wallet-react/vite.config.ts
@@ -1,14 +1,6 @@
 import react from '@vitejs/plugin-react'
-import { defineConfig } from 'vitest/config'
+import { defineConfig } from 'vite'
 
 export default defineConfig({
-  plugins: [react()],
-  test: {
-    name: 'use-wallet-react',
-    dir: './src',
-    watch: false,
-    environment: 'jsdom',
-    setupFiles: ['./vitest.setup.ts'],
-    globals: true
-  }
+  plugins: [react()]
 })