+++ packages/use-wallet-vue/vitest.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig, mergeConfig } from 'vitest/config'
+import viteConfig from './vite.config'
+
+export default mergeConfig(
+  viteConfig,
+  defineConfig({
+    test: {
+      name: 'use-wallet-vue',
+      dir: './src',
+      watch: false,
+      environment: 'jsdom',
+      setupFiles: ['./vitest.setup.ts'],
+      globals: true
+    }
+  })
+)